+++ app/src/main/java/com/daffodyl/medialibrary/models/Movie.kt
@@ -1,11 +1,11 @@
 package com.daffodyl.medialibrary.models
 
 data class Movie(
-    val id: Int,
+    val id: Long,
     val title: String,
     val format: String,
     val rating: String,
-    val runtime: Int,
+    val runtime: Long,
     val genre: String,
     val notes: String
 )
