+++ app/src/main/java/com/daffodyl/medialibrary/viewmodels/CreateVideoGameScreenViewModel.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 class CreateVideoGameScreenViewModel(
-    private val videoGameId: Int?,
+    private val videoGameId: Long?,
     private val videoGamesRepository: VideoGamesRepository
 ): ViewModel() {
     private val _title = MutableStateFlow("")
@@ -95,7 +95,7 @@ class CreateVideoGameScreenViewModel(
     }
 
     companion object {
-        var BOARD_GAME_ID_KEY = object : CreationExtras.Key<Int?> {}
+        var BOARD_GAME_ID_KEY = object : CreationExtras.Key<Long?> {}
 
         val Factory = viewModelFactory {
             initializer {
